requirements.txt
torch==1.13.1
torchaudio==0.13.1
transformers==4.26.1
pydub==0.25.1
python-multipart==0.0.6
fastapi==0.94.1
uvicorn[standard]==0.21.0
line-bot-sdk==2.4.2
marisa_trie==0.8.0
pythainlp==4.0.1
onnx
onnxruntime-gpu
optimum[onnxruntime-gpu]
